- Resolvido issue #20 e #22;  - Melhorias nas estruturas das classes (divisão de código criada a classe CommonPresenter)  - Alterado o nome da classe Presenter (do diretório do student) para ViewPresenter (dessa maneira esta compatível com a do professor em termos de nomenclatura).
