+++ jn_tester/professor/tables.py
@@ -1,6 +1,7 @@
 # -*- encoding: utf-8 -*-
 
-from .presentation import Presenter
+from .presentation import ViewPresenter
+from .export import ExportPresenter
 
 
 def view_complete_table(test_name, base_path='./', sort_by=None, presentation_format='table'):
@@ -12,7 +13,7 @@ def view_complete_table(test_name, base_path='./', sort_by=None, presentation_fo
     :param sort_by: list containing columns to be sorted.
     :param presentation_format: string of format presentation: table | text.
     :param export_format: string of format to export: csv | text."""
-    presenter = Presenter(test_name, base_path, sort_by, presentation_format)
+    presenter = ViewPresenter(test_name, base_path, sort_by, presentation_format)
     presenter.show()
 
 
@@ -24,5 +25,5 @@ def export_complete_table(test_name, base_path='./', sort_by=None, export_format
     :param sort_by: list containing columns to be sorted.
     :param presentation_format: string of format presentation: table | text.
     :param export_format: string of format to export: csv | text."""
-    presenter = Presenter(test_name, base_path, sort_by, export_format=export_format)
+    presenter = ExportPresenter(test_name, base_path, sort_by, export_format=export_format)
     presenter.export()
